Mise à jour Programme DBHELPER -> SINGLETON REMOVE De BackGroundTask Pour récuperer BD :  DBHelper dbhelper = DBHelper.getInstance(this);

+++ app/src/main/java/com/example/lamur/exchangesstudents/DBHelper.java
@@ -3,15 +3,19 @@ package com.example.lamur.exchangesstudents;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
+import android.database.MatrixCursor;
+import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
 
-import java.io.Serializable;
-import java.io.SerializablePermission;
 import java.util.ArrayList;
 
-public class DBHelper extends SQLiteOpenHelper implements Serializable {
+import static android.content.ContentValues.TAG;
+
+public class DBHelper extends SQLiteOpenHelper{
+
+    private static DBHelper sInstance;
 
     private static final int DATABASE_VERSION = 1;
     private static final String DATABASE_NAME = "Services.db";
@@ -26,6 +30,15 @@ public class DBHelper extends SQLiteOpenHelper implements Serializable {
     public static final String SERVICE_USERNAME = "type_service";
     public static final String TAUX_HORAIRE = "taux_horaires";
 
+    public static synchronized DBHelper getInstance(Context context) {
+        // Use the application context, which will ensure that you
+        // don't accidentally leak an Activity's context.
+        // See this article for more information: http://bit.ly/6LRzfx
+        if (sInstance == null) {
+            sInstance = new DBHelper(context.getApplicationContext());
+        }
+        return sInstance;
+    }
 
     public DBHelper(Context context){
 
@@ -33,14 +46,22 @@ public class DBHelper extends SQLiteOpenHelper implements Serializable {
 
     }
 
+    @Override
+    public void onConfigure(SQLiteDatabase db) {
+        super.onConfigure(db);
+       // db.setForeignKeyConstraintsEnabled(true);
+    }
+
     public void onCreate(SQLiteDatabase db){
 
-        String CREATE_PRODUCTS_TABLE = "CREATE TABLE " +
-                TABLE_USERS + "("
-                + COLUMN_ID + " INTEGER PRIMARY KEY," +
-                COLUMN_USERNAME +
-                " TEXT," + COLUMN_MDP + " INTEGER," +
-                COLUMN_ROLE + " TEXT"+ ")";
+        String CREATE_USERS_TABLE = "CREATE TABLE " + TABLE_USERS +
+                "(" +
+                COLUMN_ID + " INTEGER PRIMARY KEY," + // Define a primary key
+                COLUMN_USERNAME + " TEXT," +
+                COLUMN_MDP +  " TEXT," +
+                COLUMN_ROLE + " TEXT"+
+                ")";
+
 
         String CREATE_SERVICES_TABLE =  "CREATE TABLE " +
                 TABLE_SERVICES + "("
@@ -48,51 +69,74 @@ public class DBHelper extends SQLiteOpenHelper implements Serializable {
                 SERVICE_USERNAME +
                 " TEXT," + TAUX_HORAIRE + " DOUBLE" + ")";
 
-        db.execSQL(CREATE_PRODUCTS_TABLE);
+        db.execSQL(CREATE_USERS_TABLE);
         db.execSQL(CREATE_SERVICES_TABLE);
 
     }
 
 
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
-        onCreate(db);
+        if (oldVersion != newVersion) {
+            // Simplest implementation is to drop all old tables and recreate them
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_SERVICES);
+            onCreate(db);
+        }
+
     }
-    public void addUser(String username, String mdp, String role){
+
+    public void addOrUpdateUser(String username, String mdp, String role){
 
 
-            SQLiteDatabase db = this.getWritableDatabase();
-
+            SQLiteDatabase db = getWritableDatabase();
+          db.beginTransaction();
+          try {
             ContentValues values = new ContentValues();
             values.put(COLUMN_USERNAME, username);
             values.put(COLUMN_MDP, mdp);
             values.put(COLUMN_ROLE,role);
+              int rows = db.update(TABLE_USERS, values, COLUMN_USERNAME + "= ?", new String[]{username});
 
-            db.insert(TABLE_USERS, null, values);
-            db.close();
+              if (rows == 1) {
+                  // Get the primary key of the user we just updated
+                  String usersSelectQuery = String.format("SELECT %s FROM %s WHERE %s = ?",
+                          COLUMN_USERNAME, TABLE_USERS, COLUMN_ID);
+                  Cursor cursor = db.rawQuery(usersSelectQuery, new String[]{String.valueOf(username)});
+                  try {
+                      if (cursor.moveToFirst()) {
+                          db.setTransactionSuccessful();
+                      }
+                  } finally {
+                      if (cursor != null && !cursor.isClosed()) {
+                          cursor.close();
+                      }
+                  }
+              } else {
+                  // user with this userName did not already exist, so insert new user
+                  db.insertOrThrow(TABLE_USERS, null, values);
+                  db.setTransactionSuccessful();
+              }
+          } catch (Exception e) {
+              Log.d(TAG, "Error while trying to add or update user");
+          } finally {
+              db.endTransaction();
+          }
 
     }
 
 
 
     public boolean isReal(String username, String mdp){
-       SQLiteDatabase db = this.getReadableDatabase();
 
-        String query = "Select * FROM "
-                + TABLE_USERS
-                + " WHERE "
-                + COLUMN_USERNAME
-                + " = \""
-                + username
-                + "\""
-                + " AND "
-                + COLUMN_MDP
-                + " = \""
-                + mdp
-                + "\""
-                ;
-
-        Cursor cursor = db.rawQuery(query, null);
+        String IsReel =
+                String.format("SELECT * FROM %s WHERE %s = '%s' AND %s = '%s'",
+                        TABLE_USERS,
+                        COLUMN_USERNAME,
+                        username,
+                        COLUMN_MDP,
+                        mdp);
+        SQLiteDatabase db = getReadableDatabase();
+        Cursor cursor = db.rawQuery(IsReel, null);
 
         if(cursor.moveToFirst()){
 
@@ -110,39 +154,44 @@ public class DBHelper extends SQLiteOpenHelper implements Serializable {
     }
 
     public String infoUser(String username, String mdp) {
-        SQLiteDatabase db = this.getReadableDatabase();
+        SQLiteDatabase db = getReadableDatabase();
 
         String role = "";
-        String query = "Select role FROM "
-                + TABLE_USERS
-                + " WHERE "
-                + COLUMN_USERNAME
-                + " = \""
-                + username
-                + "\""
-                + " AND "
-                + COLUMN_MDP
-                + " = \""
-                + mdp
-                + "\"";
 
-        Cursor cursor = db.rawQuery(query, null);
 
-        if (cursor.moveToFirst()) {
-            role = cursor.getString(0);
-            cursor.close();
-            db.close();
-            return role;
+        String infoUser =
+                String.format("SELECT * FROM %s WHERE %s = '%s' AND %s = '%s'",
+                        TABLE_USERS,
+                        COLUMN_USERNAME,
+                        username,
+                        COLUMN_MDP,
+                        mdp);
 
-        } else {
-            cursor.close();
-            db.close();
-            return "";
+
+        Cursor cursor = db.rawQuery(infoUser, null);
+        try {
+            if (cursor.moveToFirst()) {
+                if (cursor.getString(cursor.getColumnIndex(COLUMN_ROLE)) != null) {
+                    role = cursor.getString(cursor.getColumnIndex(COLUMN_ROLE));
+                } else {
+                    role = "Unknown";
+                }
+            }
+
+        } catch (Exception e) {
+            Log.d(TAG, "Error while trying to get posts from database");
+        } finally {
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
         }
+
+        return  role;
     }
 
+
     public boolean deleteUser(String UserName){
-        SQLiteDatabase db = this.getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase();
         boolean result = false;
         String query = "SELECT * FROM "
                 + TABLE_USERS
@@ -203,4 +252,6 @@ public class DBHelper extends SQLiteOpenHelper implements Serializable {
         */
 
     }
+
+
 }
